Extract Pandoc SCSS handling
-> extract into standalone file
-> fully implement dark mode with handling of text highlighting
